rework the human turn systems
Previously, there was a single human turn state where our mouse event
handler did most of the work, sending events to a couple of other event
handlers to keep the code somewhat manageable.

Well even that wasn't very manageable and didn't feel right either. So
now we two states, one for when there is no marble selected and one for
when there is a marble selected.

This new design, although more complicated in the flow diagram, is much
simpler and more focused on the individual tasks of waiting for a marble
to be selected and handling clicks after a marble is selected.

NOTE: I also removed the ESC key handler for deselecting marbles since
it conflicts with the state transition (you can't transition to the same
state more than once with Bevy). I'll bring this back when I figure out
a nice way to handle it with the new design.