mark move die + click event bug fix
We're now marking the possible moves with which die the move uses so we
can figure out if there's more moves available.

There was a click event bug since both the idle and selected states have
a system that listens for click events. Maybe I should just include that
simple logic in both individually?