add buffer timer for when no moves are available
It's a bit jarring when the dice rolls, lands on the current player's
location and immediately moves to the next player's location because
there were no moves.

This change adds a buffer to the turn setup state to give the player a
second to see that they have no moves. The buffer makes this nicer but
maybe a UI element could enhance this further?