fix human destination select bug
So far I've been testing as the Blue player and no selection issues have
come up. But now that I can choose a different color (which I could have
just hard coded, but oh well) I found a bug in the destination selection
logic. The bug is related to rotating coordinates, which originally was
meant for rotating the board rather than just any set of coordinates.

Now when checking the clicked destination against the board, I'm
rotating the board coordinates and checking them against the clicked
coordinates, not the other way around.