fix (yet another) possible moves calculation bug
The logic for calculating "basic" moves (i.e., moves for marbles that
are currently on a normal board space - not the center and not the base)
had a filtering bug where I was allowing the center space as long as
the path to get there started either before the first home row space or
at the start space. This allowed the case where you could be outside the
home row and roll a large enough combined value to enter the center
space via the home row.

Now, I'm checking that the path to get to the center space does not go
through the home row at all.

I've also broken the basic move calculation logic out into a function to
help me test. I'll probably do this for the other two scenarios as well,
exiting the base and exiting the center.