simplify human turn systems
Instead of two separate states to handle the human turn, there is now
one state. The reason for two states previously was because of a shared
mouse button input handler and duplicate click events. I figured out a
better way such that the duplicate events are fine (desired, in fact).