MAJOR main menu update
This was a big one. I basically did a rewrite of the main menu so that I
could get the rules pages added in a way that was relatively simple. The
UI stuff in Bevy is definitely a pain point for me, but I think with the
way I've designed this, it actually isn't as bad.