MAJOR human turn systems rework
I wasn't happy with the current set of systems for the human turn state.
Both of the systems had a couple different responsibilities and it just
didn't feel right. Now each system has a better name and there's the
flow is better (in my opinion).

This change affected some other systems as well so you'll see those
updates here too.