change moving speed to be slightly more dynamic
The problem was that sometimes short moves looked instant and were kind
of jarring to see. With this change I've given each move a "base" speed
of 550 pixels per second. To that base speed we add the actual length
of the move (in pixels) so shorter moves are slightly slower than longer
moves and it feels much better (not perfect, but better).