fix tiny button annoyances
The main issue fixed is that buttons didn't get the correct initial
state when being created because I wasn't checking the mouse location
or state so for example if the mouse was hovering a button when it was
spawned, it wouldn't pick up the "hovered" state until you moved the
mouse.

I've also cleaned a few things with the button-related functions.