fix selected marble bug and its side effects
There was a bug where you roll doubles but have no moves, then you
select a marble and click "Done" which rolls again, but because you have
a marble still selected it is highlighted again. This didn't cause any
issues but it looked wrong.

The solution was to clear the selected marble whenever you click on
anything but that marble. However, there was a problem with this because
if you selected a move the processing systems were expecting to find
the marble selection intact. This was fixed by just moving the selected
marble into a 'moved' marble variable so that processing a move could do
its job.

There was another bug addressed with highlighting where it was possible
to re-highlight an already highlighted move:
- roll a 1
- select a marble in either the base or center
- use a double-dice power up
The fix was to double check that none of the new highlights have the
same board index as any of the current highlights.